Remove flags from Ruby shebang
I noticed in the Ruby shebang that `-K and `-U` flags were included. When I tried to run a script with the `-K` flag, I got this error:

```
ruby: warning: -K is specified; it is for 1.8 compatibility and may cause odd behavior
```

I couldn't find `-U` documented at all, which makes me think it's unused.